Add style to hero section
Add style to remove default margin on h1 so the
top of the hero heading is in line with the top
of the hero image. I realize the hero tagline
text styling is the same as the nav links, so I
combined the classes to reuse the same styling,
which required re-ordering the CSS to group the
styles together. Finally, I added style to the
sign up button to match the requirements.