Change styling on info section to align title
The info container had flexbox properties that
made it hard to center the title in the information
section. I removed the flexbox styling and
replaced the styling rules with a simpler
text-align: center on the title. I also added
padding to the title to provide the correct amount
of whitespace.