Add container div for alignment
The quote was too wide. Adding the container div
allowed the quote to conform to the same dimensions
as the width in the other sections. The citation
needed rearranging to be combined with the text
div for proper width management and whitespace
(the padding from text applies). Because of
moving the citation to get wrapped by text div,
it required extra style rules to get the citation
to look correct.