Fix alignment in information section
Information section was narrower than rest of the
page. It wasn't using the container to set the
width, so the alignment looked off compared to the
hero, quote, and CTA sections. Fixed by wrapping
the information section in the container. The style
rules changed to accommodate the container and
allowed me to get rid of the gap by using
flexbox space-between to evenly distribute the
icons.