Add width style rules to CTA container
To allow the blue box to extend past the text.
The CTA heading and text need to stay in alignment
with other content on the page. The blue box
needs to extend wider. Max-width was incorrect.
The calc allows us to add an extra 64 px on both
the left and right side of the box to keep
the text centered in the same position.